EIOState refactor and DataContext Cleanup
New previous resulting Enumerator is cahced in as a HashSet<T>, and the form is cleared once the data has been loaded into the HashSet<T>

Updated EIOState Enum, and all files that would have had the old values have also been updated.

Updated DataContext to remove a lot of the death paths.